# docs/HRM_swagger.yaml
swagger: '2.0'
info:
  version: '1.0'
  title: HRM_Project
  contact: {}
host: localhost:8081
basePath: /api
securityDefinitions:
  token:
    type: apiKey
    description: ''
    name: token
    in: header
schemes:
  - http
consumes:
  - application/json
produces:
  - application/json
paths:
  /users:
    post:
      summary: create new user
      tags:
        - User API
      operationId: createnewuser
      deprecated: false
      produces:
        - application/json
      consumes:
        - application/x-www-form-urlencoded
      parameters:
        - name: roleId
          in: formData
          required: true
          type: string
          description: ''
        - name: employeeId
          in: formData
          required: true
          type: string
          description: ''
        - name: managerId
          in: formData
          required: true
          type: string
          description: ''
        - name: firstName
          in: formData
          required: true
          type: string
          description: ''
        - name: lastName
          in: formData
          required: true
          type: string
          description: ''
        - name: email
          in: formData
          required: true
          type: string
          description: ''
        - name: phone
          in: formData
          required: true
          type: string
          description: ''
        - name: avatar
          in: formData
          required: true
          type: string
          description: ''
        - name: address
          in: formData
          required: true
          type: string
          description: ''
        - name: department
          in: formData
          required: true
          type: string
          description: ''
        - name: identificationNumber
          in: formData
          required: true
          type: integer
          format: int32
          description: ''
        - name: insuranceNumber
          in: formData
          required: true
          type: integer
          format: int32
          description: ''
      responses:
        '200':
          description: ''
          headers: {}
    get:
      summary: get all user
      tags:
        - User API
      operationId: getalluser
      deprecated: false
      produces:
        - application/json
      parameters:
        - name: currentPage
          in: query
          required: true
          type: integer
          format: int32
          description: ''
        - name: pageSize
          in: query
          required: true
          type: integer
          format: int32
          description: ''
      responses:
        '200':
          description: ''
          headers: {}
  '/users/{id}':
    delete:
      summary: delete user by id
      tags:
        - User API
      operationId: deleteuserbyid
      deprecated: false
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          required: true
          type: string
          description: ''
      responses:
        '200':
          description: ''
          headers: {}
    get:
      summary: get user by id
      tags:
        - User API
      operationId: getuserbyid
      deprecated: false
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          required: true
          type: string
          description: ''
      responses:
        '200':
          description: ''
          headers: {}
    put:
      summary: update user
      tags:
        - User API
      operationId: updateuser
      deprecated: false
      produces:
        - application/json
      consumes:
        - application/x-www-form-urlencoded
      parameters:
        - name: id
          in: path
          required: true
          type: string
          description: ''
        - name: employeeId
          in: formData
          required: true
          type: string
          description: ''
        - name: managerId
          in: formData
          required: true
          type: string
          description: ''
        - name: firstName
          in: formData
          required: true
          type: string
          description: ''
        - name: lastName
          in: formData
          required: true
          type: string
          description: ''
        - name: email
          in: formData
          required: true
          type: string
          description: ''
        - name: phone
          in: formData
          required: true
          type: integer
          format: int64
          description: ''
        - name: avatar
          in: formData
          required: true
          type: string
          description: ''
        - name: address
          in: formData
          required: true
          type: integer
          format: int32
          description: ''
        - name: department
          in: formData
          required: true
          type: string
          description: ''
        - name: identificationNumber
          in: formData
          required: true
          type: integer
          format: int32
          description: ''
        - name: insuranceNumber
          in: formData
          required: true
          type: integer
          format: int32
          description: ''
      responses:
        '200':
          description: ''
          headers: {}
...
  '/userroles/{id}':
    put:
      summary: update userrole
      tags:
        - UserRole API
      operationId: updateuserrole
      deprecated: false
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          required: true
          type: string
          description: ''
        - name: Body
          in: body
          required: true
          description: ''
          schema:
            $ref: '#/definitions/updateuserrolerequest'
      responses:
        '200':
          description: ''
          headers: {}
...
definitions:
...
  updateuserrolerequest:
    title: updateuserrolerequest
    example:
      userId: 8b8ba587-5ec4-4bac-aab7-4923f7af447e
      roleId: 258c3db8-5a98-40e4-94bc-753c2c0eb73c
    type: object
    properties:
      userId:
        type: string
      roleId:
        type: string
    required:
      - userId
      - roleId
...
security:
  - token: []
...